RAG性能调优/31-跨模态Embedding/src/requirements.txt
streamlit
pillow
dashscope
scikit-learn
numpy